refactor(planning): re-home estimateTokens and delete the dead ContextEnvelope SDK (refs #5005)
Only `estimateTokens` from `lib/orchestration/context-envelope.js` had live
callers; `buildEnvelope`, `elideEnvelope`, `envelopeToPrompt` and the
section tables were all carried as production-dead rows. The estimator now
lives in `spec-spill.js` — the module that already defined the token budget
it feeds — with `ticket-validator-sizing.js` and `checklist-threading.js`
importing it from there.
